+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -209,6 +209,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed empty MoreOptions and moved AdvancedManipulation in a dedicated location for coherency in DensityVolume.
 - Fix light limit counts specified on the HDRP asset
 - Fixed Quality Settings for SSR, Contact Shadows and Ambient Occlusion volume components
+- Unlocked SSS, SSR, Motion Vectors and Distortion frame settings for reflections probes.
 
 ### Changed
 - Color buffer pyramid is not allocated anymore if neither refraction nor distortion are enabled